requirements.txt
fastapi
uvicorn
gradio_client
python-dotenv
